Move IconTextField to Hiscore plugin
As it is the only place where it is used, it should be placed there.
