+++ runelite-client/src/main/java/net/runelite/client/plugins/hiscore/IconTextField.java
@@ -23,7 +23,7 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.ui;
+package net.runelite.client.plugins.hiscore;
 
 import java.awt.Graphics;
 import java.awt.Insets;
